refactor: split panel opening into probe, fallback, and decision
panel-open.ts had accumulated five concerns over eight rounds of chasing this
bug — capability detection, the display-mode preference, the fallback
escalation chain, context-menu registration, and i18n. Each was added where the
previous one already sat.

Split by concern:
  panel/sidepanel-probe.ts  — can this browser show a side panel?
  panel/fallback-window.ts  — opening the panel outside the side panel
  panel/context-menu.ts     — the right-click entry point
  panel-open.ts             — the decision, and the order it is made in

Not a facade: the two external callers now import from the module they actually
depend on. Registration and click handling are deliberately separated — the
context menu module knows nothing about the fallback path, which is what keeps
the dependency acyclic.

panel-open.ts is left holding only the decision, because that order is the
substance of the fix: the persisted user preference is consulted first and
outranks every probe.

Also fixes an orphaned JSDoc block: initPanelOpening's doc comment had been
left stranded above an unrelated constant by an earlier edit, documenting the
wrong symbol.

Tests stay as one suite across the three modules — what needs pinning is how
they compose, and the browser fixture that makes any of it meaningful is
shared. Behaviour is unchanged: the same 3321 tests pass, typecheck and build
clean (the one failure is the pre-existing TZ=UTC issue in prompt.test.ts).
